pyproject.toml
[project]
name = "muika-after-story"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0, <3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fleep>=1.0.1",
    "jinja2>=3.1.6",
    "mcp[cli]>=1.26.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.60.4",
    "nonebot-plugin-orm>=0.8.3",
    "nonebot-plugin-userinfo>=0.2.6",
    "nonebot2>=2.4.4",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.46",
    "watchdog>=6.0.0",
]
authors = [
    { name = "Moemu", email = "i@snowy.moe" },
]

[project.optional-dependencies]
standard = [
    "azure-ai-inference>=1.0.0b7",
    "dashscope>=1.22.1",
    "google-genai>=1.8.0",
    "ollama>=0.4.7",
    "soundfile>=0.13.1",
    "pytz>=2025.2"
]
dev = [
    "pre-commit>=4.1.0",
    "mypy>=1.15.0",
    "black>=25.1.0",
    "types-PyYAML"
]

[tool.nonebot]
adapters = [
    { name = "OneBot V12", module_name = "nonebot.adapters.onebot.v12" },
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["muika"]
builtin_plugins = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"
tag_filter = "v*"
tag_regex = '^v(?:\D*)?(?P<version>([1-9][0-9]*!)?(0|[1-9][0-9]*)(\.(0|[1-9][0-9]*))*((a|b|c|rc)(0|[1-9][0-9]*))?(\.post(0|[1-9][0-9]*))?(\.dev(0|[1-9][0-9]*))?$)$'
fallback_version = "1.0.0"

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
